backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-docx==1.1.0
pdfplumber==0.10.3
openpyxl>=3.1.0
celery==5.3.6
redis==5.0.1
sqlalchemy==2.0.25
pydantic==2.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai>=1.55.0
